# ContentApiCore/wwwroot/json/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Open Data Hub Content Api",
    "description": "Open Data Hub Content Api based on .Net Core with PostgreSQL",
    "termsOfService": "https://opendatahub.readthedocs.io/en/latest/",
    "contact": {
      "name": "Open Data Hub Team",
      "url": "https://opendatahub.com/",
      "email": "help@opendatahub.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.tourism.testingmachine.eu"
    }
  ],
  "paths": {
...
    "/v1/GeoShapes": {
      "get": {
        "tags": [
          "Geo"
        ],
        "summary": "GET GeoShapes List",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Select fields to display, More fields are indicated by separator ',' example fields=Id,Active,Shortname (default:'null' all fields are displayed). <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#fields\" target=\"_blank\">Wiki fields</a>",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "searchfilter",
            "in": "query",
            "description": "String to search for, Title in all languages are searched, (default: null) <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#searchfilter\" target=\"_blank\">Wiki searchfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawfilter",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawfilter\" target=\"_blank\">Wiki rawfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawsort",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawsort\" target=\"_blank\">Wiki rawsort</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "removenullvalues",
            "in": "query",
            "description": "Remove all Null values from json output. Useful for reducing json size. By default set to false. Documentation on <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters,-fields,-language,-searchfilter,-removenullvalues,-updatefrom#removenullvalues\" target=\"_blank\">Opendatahub Wiki</a>",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "text/csv": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "application/ld+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "application/ldjson": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              },
              "application/rawdata": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShapes"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/v1/GeoShapes/{id}": {
      "get": {
        "tags": [
          "Geo"
        ],
        "summary": "GET GeoShape Single",
        "operationId": "SingleShapes",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Tag",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Select fields to display, More fields are indicated by separator ',' example fields=Id,Active,Shortname (default:'null' all fields are displayed). <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#fields\" target=\"_blank\">Wiki fields</a>",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "removenullvalues",
            "in": "query",
            "description": "Remove all Null values from json output. Useful for reducing json size. By default set to false. Documentation on <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters,-fields,-language,-searchfilter,-removenullvalues,-updatefrom#removenullvalues\" target=\"_blank\">Opendatahub Wiki</a>",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Object created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShapes"
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoShapes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "code_rip": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "code_reg": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "code_prov": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "code_cm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "code_uts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "istatnumber": {
            "type": "string",
            "nullable": true
          },
          "abbrev": {
            "type": "string",
            "nullable": true
          },
          "type_uts": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "name_alternative": {
            "type": "string",
            "nullable": true
          },
          "shape_leng": {
            "type": "number",
            "format": "float"
          },
          "shape_area": {
            "type": "number",
            "format": "float"
          },
          "geom": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "Type": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Detail": {
            "type": "string",
            "nullable": true
          },
          "Instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          }
        }
      }
    }
  },
...
}